auto-gen-files
This branchhas the following files that introduce some new features,
1. The .mlapp file can be run to launch the app-builder and a GUI that can generate WEC-Sim Input Files. The app can also be installed and will appear under 'Apps' tab in MATLAB.  If installing the app, note, that the app files are stored in the MATLAB's Apps related data storage folder,

2. The wecsim_modelmaker.m can generate wec-sim models programmatically after an input file has been run. The code can generate bodies, ptos and constraints using the input file and the user can define the configuration of the wec-sim model. For instance, the user can define how the bodies ptos and constraints are connects whether in parallel such that all bodies are attached to the global frame or in series, such that some bodies are attached to other bodies using pto/constraint blocks while others are connected to the global frame.

Here the Postive signs to the ptoType and constraintType variables denote
parallel connections, such that the corresponding body is attached to the
global frame.
The Negative sign denotes that the body is connected to the body with
index one less than the current body.
ptoType variable defines the type of ptos such that,
1. Spherical PTO,
2. Translational PTO,
3. Translational PTO Actuation Force,
4. Translational PTO Actuation Motion,
5. Rotational PTO,
6. Rotational PTO Actuation Torque,
7. Rotational PTO Actuation Motion
constraintType varibables are defined as,
1. Fixed,
2. Spherical,
3. Translational,
4. Rotational,
5. Floating 3DOF,
6. FLoating 6DOF.

Note: The tranlational PTO ports are defined differently than all other
ptos and constraints. This code accounts for that at the end of PTO
for-loops.